requirements.txt
alabaster==0.7.7
amqp==1.4.9
anaconda-client==1.4.0
anaconda-navigator==1.1.0
anyjson==0.3.3
argcomplete==1.0.0
argh==0.26.2
astropy==1.1.2
Babel==2.2.0
backports-abc==0.4
backports.ssl-match-hostname==3.4.0.2
beautifulsoup4==4.4.1
billiard==3.3.0.23
bitarray==0.8.1
blaze==0.9.1
bokeh==0.11.1
boto==2.39.0
Bottleneck==1.0.0
cdecimal==2.3
celery==3.1.23
cffi==1.5.2
chest==0.2.3
cloudpickle==0.1.1
clyent==1.2.1
colorama==0.3.7
comtypes==1.1.2
conda-manager==0.3.1
configobj==5.0.6
cryptography==1.3
cubes==1.0.1
cycler==0.10.0
Cython==0.23.4
cytoolz==0.7.5
dask==0.8.1
datashape==0.5.1
decorator==4.0.6
dill==0.2.4
Django==1.9.6
django-admin-steroids==0.3.14
django-admin-tools==0.7.2
django-admin-tools-stats==0.6.6
django-bower==5.1.0
django-cache-utils==0.7.2
django-celery==3.1.17
django-crontab==0.7.1
django-extensions==1.6.6
django-generic-aggregation==0.4.0
django-grappelli==2.7.3
django-jsonfield==0.9.19
django-nvd3==0.9.7
django-pandas==0.4.1
django-pgviews==0.1.0
django-preserialize==1.1.0
django-pyodbc==0.3.0
django-pyodbc-azure==1.9.3.0
django-qsstats-magic==0.7.2
django-report-builder==3.3.0
django-report-utils==0.3.15
django-supervisor==0.3.4
django-taggit==0.21.2
django-taggit-serializer==0.1.5
djangorestframework==3.3.3
docutils==0.12
elasticsearch==2.4.0
elasticsearch-dsl==2.1.0
enum34==1.1.2
et-xmlfile==1.0.1
fastcache==1.0.2
Flask==0.10.1
Flask-Cors==2.1.2
funcsigs==0.4
functools32==3.2.3.post2
futures==3.0.3
geopy==1.11.0
gevent==1.1.0
greenlet==0.4.9
grin==1.2.1
h5py==2.5.0
HeapDict==1.0.0
holidays==0.4.1
idna==2.0
ipaddress==1.0.14
ipykernel==4.3.1
ipython==4.1.2
ipython-genutils==0.1.0
ipywidgets==4.1.1
itsdangerous==0.24
jdcal==1.2
jedi==0.9.0
Jinja2==2.8
JSON-Datetime==0.0.3
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.2.2
jupyter-console==4.1.1
jupyter-core==4.1.0
kombu==3.0.35
Lifetimes==0.2.0.0
llvmlite==0.9.0
locket==0.2.0
lxml==3.6.0
MarkupSafe==0.23
matplotlib==1.5.1
meld3==1.0.2
menuinst==1.3.2
mistune==0.7.2
mpmath==0.19
multipledispatch==0.4.8
nbconvert==4.1.0
nbformat==4.0.1
networkx==1.11
nltk==3.2
nose==1.3.7
notebook==4.1.0
numba==0.24.0
numexpr==2.5
numpy==1.10.4
odo==0.4.2
openpyxl==2.2.1
pandas==0.18.0
partd==0.3.2
path.py==0.0.0
pathtools==0.1.2
patsy==0.4.0
pep8==1.7.0
pickleshare==0.5
Pillow==3.1.1
ply==3.8
psutil==4.1.0
psycopg2==2.6.1
py==1.4.31
pyasn1==0.1.9
pycosat==0.6.1
pycparser==2.14
pycrypto==2.6.1
pyflakes==1.1.0
Pygments==2.1.1
pyodbc==3.0.10
pyOpenSSL==0.15.1
pyparsing==2.0.3
pyreadline==2.1
pytest==2.8.5
python-dateutil==2.4.2
python-nvd3==0.14.2
python-slugify==1.1.4
pytz==2015.7
pywin32==220
PyYAML==3.11
pyzmq==15.2.0
QtAwesome==0.3.2
qtconsole==4.2.0
QtPy==1.0
requests==2.9.1
rope==0.9.4
scikit-image==0.12.3
scikit-learn==0.17.1
scipy==0.17.0
simplegeneric==0.8.1
simplejson==3.8.2
singledispatch==3.4.0.3
six==1.10.0
snowballstemmer==1.2.1
sockjs-tornado==1.0.1
sphinx==1.3.5
sphinx-rtd-theme==0.1.9
spyder==2.3.8
SQLAlchemy==1.0.12
statsmodels==0.6.1
supervisor==3.3.1
sympy==1.0
tables==3.2.2
timelines==0.2
toolz==0.7.4
tornado==4.3
traitlets==4.2.1
unicodecsv==0.14.1
Unidecode==0.4.19
urllib3==1.16
watchdog==0.8.3
Werkzeug==0.11.4
xlrd==0.9.4
XlsxWriter==0.8.4
xlwings==0.7.0
xlwt==1.0.0
